added auth.js and counter.js store folders to break store into smaller components, completed basic redux course
