+++ src/components/Auth.js
@@ -1,28 +1,28 @@
 import classes from "./Auth.module.css";
-import { useState } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { authActions } from "../store/index";
+import { useDispatch } from "react-redux";
+import { authActions } from "../store/auth";
 
 const Auth = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
+  // const [email, setEmail] = useState("");
+  // const [password, setPassword] = useState("");
 
-  const [emailValid, setEmailIsValid] = useState(false);
-  const [passwordValid, setPasswordIsValid] = useState(false);
+  // const [emailValid, setEmailIsValid] = useState(false);
+  // const [passwordValid, setPasswordIsValid] = useState(false);
 
   const dispatch = useDispatch();
 
-  const emailChangeHandler = (event) => {
-    let emailValue = event.target.value;
-    setEmail({ email: emailValue });
-  };
+  // const emailChangeHandler = (event) => {
+  //   let emailValue = event.target.value;
+  //   setEmail({ email: emailValue });
+  // };
 
-  const passwordChangeHandler = (event) => {
-   let passwordValue = event.target.value;
-   setPassword({password: passwordValue})
-  };
+  // const passwordChangeHandler = (event) => {
+  //  let passwordValue = event.target.value;
+  //  setPassword({password: passwordValue})
+  // };
 
-  const authHandler = () => {
+  const authHandler = (event) => {
+    event.preventDefault();
     dispatch(authActions.login());
   };
 
@@ -32,14 +32,13 @@ const Auth = () => {
         <form>
           <div className={classes.control}>
             <label htmlFor="email">Email</label>
-            <input type="email" id="email" onChange={emailChangeHandler} />
+            <input type="email" id="email" />
           </div>
           <div className={classes.control}>
             <label htmlFor="password">Password</label>
             <input
               type="password"
               id="password"
-              onChange={passwordChangeHandler}
             />
           </div>
           <button onClick={authHandler}> Login</button>
@@ -48,5 +47,6 @@ const Auth = () => {
     </main>
   );
 };
+
 
 export default Auth;
