+++ src/components/Header.js
@@ -1,9 +1,20 @@
 import classes from './Header.module.css';
+import {useDispatch, useSelector} from 'react-redux'
+import { authActions } from '../store/auth';
 
 const Header = () => {
+
+  const dispatch = useDispatch();
+  const logoutHandler = (event) => {
+    event.preventDefault();
+    dispatch(authActions.logout())
+  }
+  const isAuthenticated = useSelector((state)=> state.auth.isAuthenticated)
+
   return (
     <header className={classes.header}>
       <h1>Redux Auth</h1>
+      {isAuthenticated && (
       <nav>
         <ul>
           <li>
@@ -13,10 +24,11 @@ const Header = () => {
             <a href='/'>My Sales</a>
           </li>
           <li>
-            <button>Logout</button>
+            <button onClick={logoutHandler}>Logout</button>
           </li>
         </ul>
       </nav>
+      ) }
     </header>
   );
 };
